# [Additions, Changes] Steam Achievement Downloading, Seperating Platforms Scripts
- There is now a button to force download Steam achievements, this will first delete the current ones and then download the ones from Steam. There is an extra button to make sure the user doesn't accidentally click this.
- The Steam Achievements Class I made is now in it's own seperate folder. This will allow future platforms to be included/excluded more easily if I get to that
- Steam Achievements class is now fully static, no C++ singleton pattern anymore
- Steam Callbacks Class is stored as a static TUniquePtr inside of the SteamAchievementsClass. The Callbacks class just handles callbacks
- Some smaller changes to names and types but nothing that breaks anything that existed previously

+++ Plugins/AchievementPlugin/Source/AchievementPlugin/Public/AchievementPlatformsEnum.h
@@ -13,6 +13,6 @@ UENUM()
 enum EAchievementUploadTypes : uint8
 {
 	// cannot use capitals because of the typedef conflict
-	Int32 = 0,
-	Float
+	Float = 0,
+	Int32
 };
